requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
streamlit
requests
